Makefile: clang: choose GCC_TOOLCHAIN_DIR not on LD
This causes an issue when trying to build with `make LD=ld.lld` if
ld.lld and the rest of your cross tools aren't in the same directory
(ex. /usr/local/bin) (as is the case for Android's build system), as the
GCC_TOOLCHAIN_DIR then gets set based on `which $(LD)` which will point
where LLVM tools are, not GCC/binutils tools are located.

Instead, select the GCC_TOOLCHAIN_DIR based on another tool provided by
binutils for which LLVM does not provide a substitute for, such as
elfedit.

Change-Id: Ib3b11bef571e731f9f5a25f5b77da9a0ee8476cb
Fixes commit 785f11aa595b ("kbuild: Add better clang cross build support")
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>