usb: gadget: f_gsi: Handle IPA remote wakeup notification properly
Upon bus suspend when IPA requests for usb remote wake up, it is not
allowing enough time for the host to wakeup from U3. IPA is continuously
trying to wakeup the host and eventually resets the bus. Fix this issue
by adding a flag to check the status of usb remote wakeup and ignores
the request from IPA if it is in progress.

Change-Id: Id55928b7f5d3ae7f282a3b8db44949d396db0ddd
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>