fs/f2fs: checkout latest revision of f2fs from google/bluecross
fs/f2fs: initial checkout from google/bluecross

Checked out fs/f2fs from google/bluecross

fs-verity: prototype implementation

Add the fs-verity shared code in fs/verity/, along with supporting
fields in struct inode and struct super_block.  Filesystems can hook
into this library to implement efficient file-level integrity and/or
authenticity for read-only files.  It works by implementing the
dm-verity mechanism at a file-granular level, using a Merkle tree to
verify any block in the file in log(file size) time.

fs-verity is a major new VFS feature that will take some time to get
right.  This prototype may be incompatible with the version that
eventually gets upstreamed.  If you would like to participate in the
design and implementation, please do so on the linux-fscrypt mailing
list [1].  Note that this will be a topic of discussion at LSF/MM 2018.

The Merkle tree is stored in-line with the file contents, starting at
the block past the original i_size.  A footer is stored after the Merkle
tree describing the original i_size, properties of the Merkle tree, and
other information.  The metadata is generated using a userspace tool;
then, an ioctl tells the kernel to enable fs-verity.  The metadata is
not stored in an xattr because the Merkle tree may be much larger than
XATTR_SIZE_MAX, and in the case of fs-verity combined with fscrypt it's
desirable for the metadata to be encrypted to avoid leaking plaintext
hashes.

In this prototype, an ioctl FS_IOC_SET_VERITY_MEASUREMENT tells the
kernel which measurement (hash of root hash + fs-verity footer) to
expect for a given inode, pinning it into memory.  Future work will
replace or supplement this with support for signature verification.

The i_size in the in-memory VFS inode is overridden so that userspace
only sees the original data, not the fs-verity metadata.  However, the
i_size stored on-disk remains the full i_size.

[1] http://vger.kernel.org/vger-lists.html#linux-fscrypt

Co-developed-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>

(Backported to bluecross.  Tested full series with
 'android-xfstests -c f2fs -g verity')

Conflicts:
	fs/inode.c
	include/linux/fs.h

Bug: 30972906
Bug: 67380979
Change-Id: Idab791615fe06f4f64496f94b01edbc74f4849c4

f2fs: merge miscellaneous fixes

Test: build & flash & boot with fileencryption enabled on f2fs userdata

Signed-off-by: argraur <argraur@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>