sched/fair: correctly handle cpus with same energy costs
When energy delta is same as the chosen cpu, the energy model simply
rejects the new cpu.

This is incorrect behaviour, because it ends up selecting the cpu it
visits first. It should break ties intelligently - fix it such that
it prefers cpu in a shallower idle state.

Implement this.

Change-Id: I82f39163e71037eca05ea52c69800a71b1fdcc27
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: DennySPb <dennyspb@gmail.com>