ANDROID: fix up 9p filesystem due to CFI non-upstream patches
I don't know why this suddenly stopped building, but the 9p filesystem
was not properly updated for the CFI patchset.  The change itself
(changing the type of filler_t) seems to not be fully correct, but given
that the patchset is not merged upstream yet, it's hard to determine if
that really is true or not.

This patch fixes the build, which is the main issue here.

Bug: 67506682
Cc: Sami Tolvanen <samitolvanen@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I7fcdb37998e187d06e75de5d8cb053757365b970
Signed-off-by: DennySPB <dennyspb@gmail.com>