memshare: Use carveouts for memory sharing usecases
The memory sharing usecases that memshare caters to require the memory
associated with the usecase to be XPU protected. Since the memory
is currently drawn from the common CMA pool, there is still a normal
cached mapping in the kernel for the memory. This can lead to
speculative cache-line fills before the memory is unmapped from the
stage-2 page tables of Linux, and XPU protected. Later, when the lines
are evicted from the cache (as clean-evicts), this results in a write
to the system cache, which triggers an XPU violation.

Use carveouts for the memshare usecases, to avoid creating normal
cached mappings in the kernel for the memory that the memshare
usecases need to use.

Modify allocation based on the checks for the client id of the
memshare child node.

Change-Id: I1cfdc600a1aeea06072253cbbf1564fa5bebd62b
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Signed-off-by: Manoj Prabhu B <bmanoj@codeaurora.org>
Signed-off-by: DennySPb <dennyspb@gmail.com>